pyproject.toml
[tool.poetry]
name = "azure-agent"
version = "0.1.0"
description = ""
authors = ["Arunprakaash <62831206+Arunprakaash@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.6"
azure-cli = "^2.61.0"
python-dotenv = "^1.0.1"
langgraph = "^0.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
